update(webpack): migrate to webpack@2
